refactor: separate functions in pypots.utils.metrics according to their tasks;
